move HTTPConfig to http_client module
